Moved import/export settings into workspace component
